backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.3.7
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
# sqlite3 is built into Python
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
qrcode[pil]==7.4.2
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
